Add description to README.md
A brief description of the project and the techniques employed was
written, mostly as a means for self reflection once this project
is completed. This git commit body is not really necessary - I'm only
adding it for practice purposes. Also, I don't know how to wrap this
text at 72 characters without doing it manually.